Renamed searchpicks app to searchpromotions
