--- wagtail/contrib/wagtailsearchpicks/__init__.py
@@ -1,1 +1,0 @@
-default_app_config = 'wagtail.contrib.wagtailsearchpicks.apps.WagtailSearchPicksAppConfig'
