--- wagtail/contrib/wagtailsearchpicks/apps.py
@@ -1,7 +1,0 @@
-from django.apps import AppConfig
-
-
-class WagtailSearchPicksAppConfig(AppConfig):
-    name = 'wagtail.contrib.wagtailsearchpicks'
-    label = 'wagtailsearchpicks'
-    verbose_name = "Wagtail search picks"
