+++ wagtail/contrib/wagtailsearchpromotions/__init__.py
@@ -1,0 +1,1 @@
+default_app_config = 'wagtail.contrib.wagtailsearchpromotions.apps.WagtailSearchPromotionsAppConfig'
