+++ wagtail/contrib/wagtailsearchpromotions/apps.py
@@ -1,0 +1,7 @@
+from django.apps import AppConfig
+
+
+class WagtailSearchPromotionsAppConfig(AppConfig):
+    name = 'wagtail.contrib.wagtailsearchpromotions'
+    label = 'wagtailsearchpromotions'
+    verbose_name = "Wagtail search promotions"
