+++ wagtail/contrib/wagtailsearchpromotions/forms.py
@@ -4,7 +4,7 @@ from django.utils.translation import ugettext_lazy as _
 
 from wagtail.wagtailadmin.widgets import AdminPageChooser
 from wagtail.wagtailsearch.models import Query
-from wagtail.contrib.wagtailsearchpicks.models import SearchPromotion
+from wagtail.contrib.wagtailsearchpromotions.models import SearchPromotion
 
 
 class SearchPromotionForm(forms.ModelForm):
