+++ wagtail/contrib/wagtailsearchpromotions/models.py
