+++ wagtail/contrib/wagtailsearchpromotions/templates/wagtailsearchpromotions/add.html
