+++ wagtail/contrib/wagtailsearchpromotions/templates/wagtailsearchpromotions/confirm_delete.html
