+++ wagtail/contrib/wagtailsearchpromotions/templates/wagtailsearchpromotions/edit.html
