+++ wagtail/contrib/wagtailsearchpromotions/templates/wagtailsearchpromotions/includes/searchpromotion_form.html
