+++ wagtail/contrib/wagtailsearchpromotions/templates/wagtailsearchpromotions/index.html
