+++ wagtail/contrib/wagtailsearchpromotions/templates/wagtailsearchpromotions/list.html
