+++ wagtail/contrib/wagtailsearchpromotions/templates/wagtailsearchpromotions/results.html
