+++ wagtail/contrib/wagtailsearchpromotions/templatetags/wagtailsearchpromotions_tags.py
